requirements.txt
yara-python==4.2.3
psutil==5.9.5
requests==2.31.0
scapy==2.5.0
tqdm==4.66.1
pyinotify==0.9.6
scikit-learn==1.3.2
pandas==2.1.1
numpy==1.26.0
